requirements.txt
streamlit==0.77.0
pandas==1.0.5
matplotlib==3.2.2
numpy==1.18.5
xlrd==1.2.0
altair==4.1.0
Pillow==8.1.2
